migrate slam map to blocks
