+++ web/frontend/package.json
@@ -75,5 +75,10 @@
     "cypress:ci": "cypress run",
     "cypress": "cypress open",
     "test:unit": "echo \"no tests exist yet\""
+  },
+  "dependencies": {
+    "@floating-ui/dom": "^1.5.3",
+    "@viamrobotics/prime-blocks": "^0.0.7",
+    "@viamrobotics/prime-core": "^0.0.43"
   }
 }
