--- web/frontend/src/components/slam/2d/constants.ts
@@ -1,7 +1,0 @@
-export const renderOrder = {
-  grid: 0,
-  axes: 1,
-  points: 2,
-  motionPath: 3,
-  markers: 4,
-} as const;
