--- web/frontend/src/components/slam/2d/legend.svelte
@@ -1,69 +1,0 @@
-<!--
-  @component
-  A user legend for the SLAM map axes.
--->
-<svg
-  class="Axes-Legend absolute right-3 top-3"
-  width="30"
-  height="30"
-  viewBox="0 0 30 30"
-  fill="none"
->
-  <rect
-    width="10"
-    height="10"
-    rx="5"
-    fill="green"
-  />
-  <path
-    d="M4.66278 6.032H4.51878L2.76678 2.4H3.51878L4.95078 5.456H5.04678L6.47878
-            2.4H7.23078L5.47878 6.032H5.33478V8H4.66278V6.032Z"
-    fill="white"
-  />
-
-  <rect
-    x="20"
-    y="20"
-    width="10"
-    height="10"
-    rx="5"
-    fill="red"
-  />
-  <path
-    d="M23.6708 22.4L24.9268 24.88H25.0708L26.3268 22.4H27.0628L25.6628 25.144V25.24L27.0628
-            28H26.3268L25.0708 25.504H24.9268L23.6708 28H22.9348L24.3348 25.24V25.144L22.9348 22.4H23.6708Z"
-    fill="white"
-  />
-
-  <rect
-    x="4"
-    y="9"
-    width="2"
-    height="17"
-    fill="green"
-  />
-  <rect
-    x="21"
-    y="24"
-    width="2"
-    height="17"
-    transform="rotate(90 21 24)"
-    fill="red"
-  />
-  <rect
-    x="0.5"
-    y="20.5"
-    width="9"
-    height="9"
-    rx="4.5"
-    fill="#E0FAE3"
-  />
-  <rect
-    x="0.5"
-    y="20.5"
-    width="9"
-    height="9"
-    rx="4.5"
-    stroke="#3D7D3F"
-  />
-</svg>
