+++ web/frontend/src/components/slam/index.svelte
@@ -5,6 +5,7 @@
   import { onMount, onDestroy } from 'svelte';
 
   import { SlamClient, type Pose, type ServiceError } from '@viamrobotics/sdk';
+  import { SlamMap2D } from '@viamrobotics/prime-blocks';
   import { copyToClipboard } from '@/lib/copy-to-clipboard';
   import { filterSubtype } from '@/lib/resource';
   import { moveOnMap, stopMoveOnMap } from '@/api/motion';
@@ -13,7 +14,7 @@
   import { components } from '@/stores/resources';
   import Collapse from '@/lib/components/collapse.svelte';
   import PCD from '@/components/pcd/pcd-view.svelte';
-  import Slam2D from './2d/index.svelte';
+  import Dropzone from '@/lib/components/dropzone.svelte';
   import { useRobotClient, useDisconnect } from '@/hooks/robot-client';
   import type { SLAMOverrides } from '@/types/overrides';
   import { rcLogConditionally } from '@/lib/log';
@@ -51,6 +52,7 @@
   let durationInterval: number | undefined;
   let newMapName = '';
   let mapNameError = '';
+  let motionPath: string | undefined;
 
   $: loaded2d = pointcloud !== undefined && pose !== undefined;
   $: moveClicked = $operations.find(({ op }) =>
@@ -371,6 +373,10 @@
     newMapName = event.detail.value;
     mapNameError = overrides?.validateMapName(newMapName) || '';
   };
+
+  const handleDrop = (event: CustomEvent<string>) => {
+    motionPath = event.detail;
+  };
 </script>
 
 <Collapse title={name} on:toggle={toggleExpand}>
@@ -565,7 +571,7 @@
           <div class="flex flex-row pl-5 py-2 border-b border-b-light">
             <div class="flex flex-col gap-0.5">
               <div class="flex gap-2">
-                <p class="text-xs">Current position</p>
+                <p class="text-xs">Current base position</p>
                 <button
                   class="text-xs hover:underline"
                   on:click={() =>
@@ -589,7 +595,7 @@
               {/if}
             </div>
             <div class="flex flex-col gap-0.5 pl-10">
-              <p class="text-xs">Current orientation</p>
+              <p class="text-xs">Current base orientation</p>
 
               {#if pose}
                 <div class="flex flex-row items-center">
@@ -618,13 +624,22 @@
             />
           </div>
 
-          <Slam2D
-            {pointcloud}
-            {pose}
-            {destination}
-            helpers={showAxes}
-            on:click={handle2dRenderClick}
-          />
+          <Dropzone on:drop={handleDrop}>
+            <div class="relative w-full h-[400px]">
+              <SlamMap2D
+                {pointcloud}
+                {destination}
+                {motionPath}
+                basePose={pose ? {
+                  x: pose.x,
+                  y: pose.y,
+                  theta: pose.theta,
+                } : undefined}
+                helpers={showAxes}
+                on:click={handle2dRenderClick}
+              />
+            </div>
+          </Dropzone>
         </div>
       {/if}
     </div>
